Move episodes command to the domain layer
