+++ lib/pod/commands/episodes/output.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module Pod
-  module Outputs
-    module Text
-      class Episodes < ::Pod::Commands::BaseOutput
+  module Commands
+    module Episodes
+      class Output < ::Pod::Commands::BaseOutput
         def call
           case @context[:details]
           when :records_found
