+++ lib/pod/outputs/text.rb
@@ -2,7 +2,6 @@
 
 require_relative "text/init"
 require_relative "text/podcasts"
-require_relative "text/episodes"
 require_relative "text/open"
 require_relative "text/sync"
 require_relative "text/update"
