+++ spec/pod/commands/episodes/runner_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative "../../support/test_helpers"
+require_relative "../../../support/test_helpers"
 
-RSpec.describe Pod::Commands::Episodes do
+RSpec.describe Pod::Commands::Episodes::Runner do
   describe  "#call", :init_pod do
     context "when there are episodes", :populate_db do
       it "returns a success response with the table records" do
